More significant progress

+++ src/Linear/Runtime/Deserializers/PrimitiveDeserializer.cs
@@ -25,7 +25,7 @@ namespace Linear.Runtime.Deserializers
 
         /// <inheritdoc />
         public object Deserialize(StructureInstance instance, Stream stream, byte[] tempBuffer, long offset, bool littleEndian,
-            Dictionary<string, object>? parameters, int length = 0)
+            Dictionary<string, object>? parameters, long length = 0)
         {
             return Type.GetTypeCode(_type) switch
             {
